Add db frontend (#36)
* backend: Phase A - widen box coords, bulk pages endpoint, seed notebook

* frontend: Phase B - add notebooksApi and pagesApi clients

* frontend: Phase C - rewrite notesStore for server sync with async actions and explicit save

* frontend: Phase D - wire explicit save calls at commit points (blur/Enter, drag-end, route-leave, manual button)

* backend: add defensive null checks and error handling to applyBoxes/applyReferences

* backend: add setId() to TextBox/TextReference; assign IDs from payload in applyBoxes/applyReferences

* backend: fix 500 on page save - use saveAndFlush and let Hibernate generate IDs for new boxes/references instead of forcing client IDs (avoids detached-entity cascade error)

---------

+++ backend/src/main/java/com/feynman/backend/dto/TextBoxPayload.java
@@ -5,9 +5,9 @@ import java.util.List;
 /** A free-floating text box on a page's canvas. */
 public record TextBoxPayload(
         String id,
-        int x,
-        int y,
-        Integer width,
+        double x,
+        double y,
+        Double width,
         String text,
         List<TextReferencePayload> references
 ) {
